refactor: rename "lib" to "common"
